ai-model/requirements.txt
# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# NLP
nltk>=3.6.2
spacy>=3.1.0

# ML Frameworks
xgboost>=1.4.2
lightgbm>=3.3.1
tensorflow>=2.6.0

# Utility Libraries
joblib>=1.0.1
argparse>=1.4.0
python-dotenv>=0.19.0
web3>=5.24.0
sqlalchemy>=1.4.23
tqdm>=4.62.0

# API & Deployment
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
requests>=2.26.0
python-multipart>=0.0.5

# Database
psycopg2-binary>=2.9.0

# Logging & Tools
tqdm==4.62.3
pillow==9.0.1
opencv-python==4.5.5.64

# Data Visualization
plotly>=5.3.1
dash>=2.0.0
dash-bootstrap-components>=1.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
matplotlib>=3.4.3
seaborn>=0.11.2
